Update `nvidia-glx-driver` to 384.59
This change introduces new splitting into the package, so that the main
package now depends on the `nvidia-glx-driver-common` subpackage. This has
enabled us to split out a new `nvidia-glx-driver-current` subpackage which
now contains modules for `linux-current` (i.e. 4.12)

The main package will continue to provide for linux-lts to ensure a consistent
upgrade path, and the new package will enable NVIDIA usage on 4.12 kernels.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>